requirements.txt
Flask>=3.1.0,<4.0.0
Flask-SQLAlchemy>=3.1.0,<4.0.0
Flask-Login>=0.6.0,<0.7.0
Flask-WTF>=1.2.0,<2.0.0
Flask-Migrate>=4.0.0,<5.0.0
email_validator>=2.2.0,<3.0.0
python-dotenv>=1.1.0,<2.0.0
Werkzeug>=3.1.0,<4.0.0
Jinja2>=3.1.0,<4.0.0
SQLAlchemy>=2.0.0,<3.0.0
WTForms>=3.2.0,<4.0.0
gunicorn>=21.2.0,<22.0.0
xhtml2pdf>=0.2.11,<0.3.0
openpyxl>=3.1.0,<4.0.0
